# Remove supplyChanged()